requirements.txt
flask==3.0.3
apscheduler==3.10.4
anthropic==0.28.0
python-docx==1.1.2
python-dotenv==1.0.1
langdetect==1.0.9
requests==2.32.3
beautifulsoup4==4.12.3
gunicorn==21.2.0
